エンドポイントのファイルを分割&削除ボタン追加
