+++ app/routes/items/action.ts
@@ -1,0 +1,16 @@
+import { type ClientActionFunction } from "@remix-run/react";
+
+import { addItem } from "./actions/addItem";
+
+export const action: ClientActionFunction = async (args) => {
+  const { request } = args;
+  switch (request.method) {
+    case "POST":
+      return addItem(args);
+    // case "DELETE":
+    //   return deleteItem(args);
+    default:
+      console.error("Invalid method", request.method);
+      return new Response(null, { status: 405 });
+  }
+};
